add NodeFile
