+++ src/managers/InMemoryHistoryManager.java
@@ -85,14 +85,4 @@ public class InMemoryHistoryManager implements HistoryManager {
     }
 }
 
-class Node {
-    public Task data;
-    public Node next;
-    public Node prev;
 
-    public Node(Task data) {
-        this.data = data;
-        this.next = null;
-        this.prev = null;
-    }
-}
