+++ src/managers/Node.java
@@ -1,0 +1,15 @@
+package managers;
+
+import classes.tasks.Task;
+
+class Node {
+    public Task data;
+    public Node next;
+    public Node prev;
+
+    public Node(Task data) {
+        this.data = data;
+        this.next = null;
+        this.prev = null;
+    }
+}
